ci: Add Docker build and test job to code-test workflow (#66)
# Pull Request

## Description

This change adds a new job called `docker-test` to the GitHub Actions
workflow. The job builds and runs a Docker image to test the GitHub
Stats Analyser.

The new job includes the following steps:
1. Checkout the repository
2. Set up Docker Buildx
3. Build the Docker image using the `just docker-build` command
4. Run the Docker image with a sample repository owner

This addition ensures that the Docker image can be successfully built
and executed as part of the continuous integration process, providing an
additional layer of testing for the project.

fixes #64