ci: Fix SonarCloud coverage path and update test commands (#70)
# Pull Request

## Description

This change enhances the code coverage reporting and SonarCloud
integration in the project:

1. Updated the `unit-test` and `unit-test-debug` commands in the
Justfile to use `--cov=.` instead of `--cov=analyser`. This modification
ensures that code coverage is calculated for the entire project rather
than just the `analyser` directory.

2. Added a new step in the GitHub Actions workflow (`code-test.yml`) to
override the coverage source path for SonarCloud. This step uses `sed`
to replace the source path in the `coverage.xml` file, ensuring that
SonarCloud can correctly map the coverage data to the source code.

These changes aim to improve the accuracy and completeness of code
coverage reporting, particularly when integrating with SonarCloud for
code quality analysis.

fixes #69