chore: Remove version from pyproject.toml (#77)
# Pull Request

## Description

This change removes the version specification from the `pyproject.toml`
file. The `version = "0.1.0"` line has been deleted, which allows for
more flexible version management of the project. This modification can
be beneficial when using automated versioning tools or when the version
is managed through other means, such as git tags or CI/CD pipelines.

fixes #78