chore: Add setup-dependencies directory to Dependabot config (#133)
# Pull Request

## Description

This change updates the Dependabot configuration to monitor GitHub
Actions dependencies in an additional directory. Specifically, it adds
the ".github/actions/setup-dependencies" directory to the list of
monitored paths. This ensures that Dependabot will check for updates in
both the root directory and the setup-dependencies action directory,
potentially improving the maintenance and security of the project's
GitHub Actions.

fixes #134