chore: Fix Docker image reference in action.yml (#149)
# Pull Request

## Description

This change updates the `action.yml` file to use the correct syntax for
specifying a Docker image in GitHub Actions. The `uses` key has been
replaced with `image` to properly reference the Docker image for the
GitHub Stats Analyser action. This modification ensures that the action
will correctly utilise the specified Docker image when running.

fixes #150