ci: Add test action and output validation (#147)
# Pull Request

## Description

This change adds a new job called "test-action" to the code-test.yml
workflow. This job runs on Ubuntu and performs the following steps:

1. Checks out the repository
2. Runs the GitHub Stats Analyser using the action defined in the
current repository
3. Executes a new test script (test_output.sh) to validate the action's
output

The new test_output.sh script is also added, which checks for the
existence and non-emptiness of the repository_statistics.json file
generated by the action.

These additions enhance the testing process by ensuring that the action
runs successfully and produces the expected output file.

fixes #148