chore: Add version check workflow and version file (#145)
# Pull Request

## Description

This change introduces a new GitHub Actions workflow named "Version
Check" in the `.github/workflows/version-check.yml` file. The workflow
is designed to run on push events to the main branch and various pull
request events.

The workflow performs a version check using the `issue-ops/semver@v2`
action, which verifies the semantic versioning in the `action.yml` file.
It's configured to run in check-only mode, meaning it will not make any
changes but will report on version discrepancies.

The workflow is set to run on Ubuntu latest and excludes
Dependabot-initiated actions. It has specific permissions set for
checks, contents, and pull requests to ensure proper functionality
within the GitHub environment.

This addition helps maintain consistent versioning practices across the
project and can catch potential version-related issues early in the
development process.

fixes #146